Automating gain AudioParam with linearRampToValueAtTime introduces buzzing distortion https://bugs.webkit.org/show_bug.cgi?id=100885
Reviewed by Kenneth Russell.

AudioParamTimeline needs to use double-precision for time-values to avoid drift and precision issues.

Covered by existing tests.

* Modules/webaudio/AudioParam.cpp:
(WebCore::AudioParam::calculateTimelineValues):
* Modules/webaudio/AudioParamTimeline.cpp:
(WebCore::AudioParamTimeline::valueForContextTime):
(WebCore::AudioParamTimeline::valuesForTimeRange):
(WebCore::AudioParamTimeline::valuesForTimeRangeImpl):
* Modules/webaudio/AudioParamTimeline.h:
(AudioParamTimeline):

Canonical link: https://commits.webkit.org/119210@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@133365 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/Modules/webaudio/AudioParamTimeline.h
@@ -60,7 +60,7 @@ public:
     // controlRate is the rate (number per second) at which parameter values will be calculated.
     // It should equal sampleRate for sample-accurate parameter changes, and otherwise will usually match
     // the render quantum size such that the parameter value changes once per render quantum.
-    float valuesForTimeRange(float startTime, float endTime, float defaultValue, float* values, unsigned numberOfValues, float sampleRate, float controlRate);
+    float valuesForTimeRange(double startTime, double endTime, float defaultValue, float* values, unsigned numberOfValues, double sampleRate, double controlRate);
 
     bool hasValues() { return m_events.size(); }
 
@@ -103,7 +103,7 @@ private:
     };
 
     void insertEvent(const ParamEvent&);
-    float valuesForTimeRangeImpl(float startTime, float endTime, float defaultValue, float* values, unsigned numberOfValues, float sampleRate, float controlRate);
+    float valuesForTimeRangeImpl(double startTime, double endTime, float defaultValue, float* values, unsigned numberOfValues, double sampleRate, double controlRate);
 
     Vector<ParamEvent> m_events;
